requirements.txt
Jinja2-2.10
MarkupSafe-1.0
Werkzeug-0.14.1
click-6.7
flask-0.12.2
itsdangerous-0.24